+++ routes/admin.php
@@ -37,22 +37,6 @@ Route::middleware([Kernel::ADMIN_ACCESS, Kernel::JSON_API])->group(function () {
         'campaigns/bid-strategy/media/{medium}/uuid-default',
         [BidStrategyController::class, 'patchBidStrategyUuidDefault']
     );
-
-    Route::post('users/{id}/switchToModerator', [AdminController::class, 'switchUserToModerator']);
-    Route::post('users/{id}/ban', [AdminController::class, 'banUser']);
-    Route::post('users/{id}/unban', [AdminController::class, 'unbanUser']);
-    Route::post('users/{id}/delete', [AdminController::class, 'deleteUser']);
-});
-
-Route::middleware([Kernel::MODERATOR_ACCESS, Kernel::JSON_API])->group(function () {
-    Route::post('users/{id}/confirm', [AuthController::class, 'confirm']);
-
-    Route::post('users/{id}/switchToAgency', [AdminController::class, 'switchUserToAgency']);
-    Route::post('users/{id}/switchToRegular', [AdminController::class, 'switchUserToRegular']);
-    Route::post('users/{id}/grantAdvertising', [AdminController::class, 'grantAdvertising']);
-    Route::post('users/{id}/denyAdvertising', [AdminController::class, 'denyAdvertising']);
-    Route::post('users/{id}/grantPublishing', [AdminController::class, 'grantPublishing']);
-    Route::post('users/{id}/denyPublishing', [AdminController::class, 'denyPublishing']);
 });
 
 Route::middleware([Kernel::AGENCY_ACCESS, Kernel::JSON_API])->group(function () {