+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Metaverse zones consolidation
 - Send medium and vendor to AdPay
 - Campaign amounts in USD (APIv2)
+### Fixed
+- Missing meta in event pagination
 
 ## [2.2.2] - 2022-12-12
 ### Added