+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Metaverse zones consolidation
 - Send medium and vendor to AdPay
 - Campaign amounts in USD (APIv2)
+- Legacy user actions which are handled by monitoring API
 ### Fixed
 - Missing meta in event pagination