+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.2.3] - 2022-12-19
 ### Added
 - Zone scopes
 ## Changed
@@ -769,7 +771,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v2.2.1...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v2.2.3...develop
+[2.2.3]: https://github.com/adshares/adserver/compare/v2.2.2...v2.2.3
 [2.2.2]: https://github.com/adshares/adserver/compare/v2.2.1...v2.2.2
 [2.2.1]: https://github.com/adshares/adserver/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/adshares/adserver/compare/v2.1.1...v2.2.0